Object is moved. Server will be killed if Marta is not learning
