+++ lib/object.rb
